Implement non-blk optical switch with GUI
The basic function of a 4X4 non-blocking optical switch implemanted using Python with a GUI.
The searching algorithm is based on exhaustive approach and needs to be optimized in the future.
Moreover, the current GUI is quite basic and can be improved for better user experience. The AddNode button is currently non-functional.